Fix templates styling.

+++ client/utils/apiWrapper.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import fetch from "isomorphic-unfetch";
+import { checkValidUser } from "./validator";
 
 const BASE_URL = process.env.BACKEND_URL ?? "http://localhost:5000/api";
 // const BASE_URL = process.env.BACKEND_URL ?? "http://52.240.158.249:5000/api"; // leave this in, this is Arpan's url
@@ -17,6 +18,7 @@ export const getModels = (sdg_query, searchPage = null) => {
   } else {
     requestString = `${BASE_URL}/models`;
   }
+  console.log(requestString);
   return axios
     .get(requestString, {
       headers: {
@@ -49,80 +51,92 @@ export const getModelsByID = Model_ID => {
       });
     });
 };
-export const getUpdates = (numUpdates, currentIndex) => {
+export const getUpdates = async (numUpdates, currentIndex) => {
   /**
    * Returns min(available updates, numUpdates) projects with ID greater than last_id query
    * Returns GET_UPDATES_FAIL upon failure
    */
-  const requestString = `${BASE_URL}/user/updates/${numUpdates}/${currentIndex}`;
-  return axios
-    .get(requestString, {
-      headers: {
-        "Content-Type": "application/JSON",
-        "x-access-token": localStorage.getItem("token")
-      }
-    })
-    .catch(error => {
-      ({
-        type: "GET_UPDATES_FAIL",
-        error
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/user/updates/${numUpdates}/${currentIndex}`;
+    return axios
+      .get(requestString, {
+        headers: {
+          "Content-Type": "application/JSON",
+          "x-access-token": localStorage.getItem("token")
+        }
+      })
+      .catch(error => {
+        ({
+          type: "GET_UPDATES_FAIL",
+          error
+        });
       });
-    });
+  }
 };
-export const addModel = data => {
+export const addModel = async data => {
   /**
    * Adds a model
    * Returns POST_MODEL_FAIL upon failure
    */
-  const requestString = `${BASE_URL}/models`;
-  return axios
-    .post(requestString, data, {
-      headers: {
-        "Content-Type": "application/JSON"
-      }
-    })
-    .catch(error => {
-      ({
-        type: "POST_MODEL_FAIL",
-        error
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/models`;
+    return axios
+      .post(requestString, data, {
+        headers: {
+          "Content-Type": "application/JSON"
+        }
+      })
+      .catch(error => {
+        ({
+          type: "POST_MODEL_FAIL",
+          error
+        });
       });
-    });
+  }
 };
-export const editModel = (data, Model_ID) => {
+export const editModel = async (data, Model_ID) => {
   /**
    * Edits a model
    * Returns UPDATE_MODEL_FAIL upon failure
    */
-  const requestString = `${BASE_URL}/models/${Model_ID}`;
-  return axios
-    .put(requestString, data, {
-      headers: {
-        "Content-Type": "application/JSON"
-      }
-    })
-    .catch(error => {
-      ({
-        type: "UPDATE_MODEL_FAIL",
-        error
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/models/${Model_ID}`;
+    return axios
+      .put(requestString, data, {
+        headers: {
+          "Content-Type": "application/JSON"
+        }
+      })
+      .catch(error => {
+        ({
+          type: "UPDATE_MODEL_FAIL",
+          error
+        });
       });
-    });
+  }
 };
-export const deleteForm = Model_ID => {
+export const deleteForm = async Model_ID => {
   /**
    * Deletes a model
    * Returns DELETE_MODEL_FAIL upon failure
    */
-  const requestString = `${BASE_URL}/models/${Model_ID}`;
-  return axios
-    .delete(requestString, {
-      headers: {}
-    })
-    .catch(error => {
-      ({
-        type: "DELETE_MODEL_FAIL",
-        error
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/models/${Model_ID}`;
+    return axios
+      .delete(requestString, {
+        headers: {}
+      })
+      .catch(error => {
+        ({
+          type: "DELETE_MODEL_FAIL",
+          error
+        });
       });
-    });
+  }
 };
 
 export const register = (
@@ -359,44 +373,52 @@ export const resendPIN = () => {
   }
 };
 
-export const userInfo = () => {
-  try {
-    return fetch(`${BASE_URL}/auth/getUser`, {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        token: localStorage.getItem("token"),
-        google: localStorage.getItem("google") ? true : false
-      }
-    });
-  } catch (err) {
-    return err;
+export const userInfo = async () => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    try {
+      return fetch(`${BASE_URL}/auth/getUser`, {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          token: localStorage.getItem("token"),
+          google: localStorage.getItem("google") ? true : false
+        }
+      });
+    } catch (err) {
+      return err;
+    }
   }
 };
 
-export const saveDescription = (
+export const saveDescription = async (
   model_id,
   phaseName,
   stageName,
   description,
   subDescription
 ) => {
-  const requestString = `${BASE_URL}/models/${model_id}/${phaseName}/${stageName}/description`;
-  return axios
-    .put(
-      requestString,
-      { description, subDescription },
-      {
-        headers: {
-          "Content-Type": "application/JSON",
-          "x-access-token": localStorage.getItem("token")
+  const validUser = await checkValidUser();
+  if (validUser) {
+    console.log("SHOULD NOT SEE THIS");
+    const requestString = `${BASE_URL}/models/${model_id}/${phaseName}/${stageName}/description`;
+    console.log(requestString);
+    return axios
+      .put(
+        requestString,
+        { description, subDescription },
+        {
+          headers: {
+            "Content-Type": "application/JSON",
+            "x-access-token": localStorage.getItem("token")
+          }
         }
-      }
-    )
-    .catch(error => ({
-      type: "SAVE_DESCRIPTION_FAIL",
-      error
-    }));
+      )
+      .catch(error => ({
+        type: "SAVE_DESCRIPTION_FAIL",
+        error
+      }));
+  }
 };
 
 export const getDescription = (model_id, phaseName, stageName) => {
@@ -435,41 +457,48 @@ export const createUser = newUser => {
   }
 };
 
-export const updateUser = updatedUser => {
-  try {
-    return (
-      fetch(`${BASE_URL}/users/userInfo`),
-      {
-        method: "PUT",
+export const updateUser = async updatedUser => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    try {
+      return (
+        fetch(`${BASE_URL}/users/userInfo`),
+        {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+            "x-access-token": localStorage.getItem("token")
+          },
+          body: JSON.stringify(updatedUser)
+        }
+      );
+    } catch (err) {
+      return err;
+    }
+  }
+};
+
+export const deleteUser = async () => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    try {
+      return fetch(`${BASE_URL}/users/userInfo`, {
+        method: "DELETE",
         headers: {
           "Content-Type": "application/json",
           "x-access-token": localStorage.getItem("token")
-        },
-        body: JSON.stringify(updatedUser)
-      }
-    );
-  } catch (err) {
-    return err;
+        }
+      });
+    } catch (err) {
+      return err;
+    }
   }
 };
 
-export const deleteUser = () => {
+export const checkAdminPrivilege = () => {
+  const requestString = `${BASE_URL}/users/`;
   try {
-    return fetch(`${BASE_URL}/users/userInfo`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        "x-access-token": localStorage.getItem("token")
-      }
-    });
-  } catch (err) {
-    return err;
-  }
-};
-
-export const getFollowingProjects = () => {
-  try {
-    return fetch(`${BASE_URL}/users/followingProjects`, {
+    return fetch(requestString, {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
@@ -480,80 +509,108 @@ export const getFollowingProjects = () => {
     return err;
   }
 };
-
-export const followProject = projId => {
-  try {
-    return fetch(`${BASE_URL}/users/followingProjects`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-        "x-access-token": localStorage.getItem("token")
-      },
-      body: JSON.stringify({ projId })
-    });
-  } catch (err) {
-    return err;
+export const getFollowingProjects = async () => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    try {
+      return fetch(`${BASE_URL}/users/followingProjects`, {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+          "x-access-token": localStorage.getItem("token")
+        }
+      });
+    } catch (err) {
+      return err;
+    }
   }
 };
 
-export const unfollowProject = projId => {
-  try {
-    return fetch(`${BASE_URL}/users/followingProjects`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        "x-access-token": localStorage.getItem("token")
-      },
-      body: JSON.stringify({ projId })
-    });
-  } catch (err) {
-    return err;
+export const followProject = async projId => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    try {
+      return fetch(`${BASE_URL}/users/followingProjects`, {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+          "x-access-token": localStorage.getItem("token")
+        },
+        body: JSON.stringify({ projId })
+      });
+    } catch (err) {
+      return err;
+    }
   }
 };
 
-export const postComment = (model_id, commentBody) => {
-  const requestString = `${BASE_URL}/comment`;
-  return axios
-    .post(
-      requestString,
-      {
-        commentLocation: model_id,
-        comment: commentBody
-      },
-      {
+export const unfollowProject = async projId => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    try {
+      return fetch(`${BASE_URL}/users/followingProjects`, {
+        method: "DELETE",
         headers: {
-          "Content-Type": "application/JSON",
+          "Content-Type": "application/json",
           "x-access-token": localStorage.getItem("token")
-        }
-      }
-    )
-    .catch(error => ({
-      type: "SAVE_COMMENT_FAIL",
-      error
-    }));
+        },
+        body: JSON.stringify({ projId })
+      });
+    } catch (err) {
+      return err;
+    }
+  }
 };
 
-export const postCommentThread = (model_id, parentIndex, commentBody) => {
-  const requestString = `${BASE_URL}/comment/thread`;
-  return axios
-    .post(
-      requestString,
-      {
-        commentLocation: `${model_id}`,
-        commentIndex: parentIndex,
-        comment: commentBody
-      },
-      {
-        headers: {
-          "Content-Type": "application/JSON",
-          "x-access-token": localStorage.getItem("token")
+export const postComment = async (model_id, commentBody) => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/comment`;
+    return axios
+      .post(
+        requestString,
+        {
+          commentLocation: model_id,
+          comment: commentBody
+        },
+        {
+          headers: {
+            "Content-Type": "application/JSON",
+            "x-access-token": localStorage.getItem("token")
+          }
         }
-      }
-    )
-    .catch(error => ({
-      type: "SAVE_COMMENT_FAIL",
-      error
-    }));
+      )
+      .catch(error => ({
+        type: "SAVE_COMMENT_FAIL",
+        error
+      }));
+  }
+};
+
+export const postCommentThread = async (model_id, parentIndex, commentBody) => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/comment/thread`;
+    return axios
+      .post(
+        requestString,
+        {
+          commentLocation: `${model_id}`,
+          commentIndex: parentIndex,
+          comment: commentBody
+        },
+        {
+          headers: {
+            "Content-Type": "application/JSON",
+            "x-access-token": localStorage.getItem("token")
+          }
+        }
+      )
+      .catch(error => ({
+        type: "SAVE_COMMENT_FAIL",
+        error
+      }));
+  }
 };
 
 export const getComments = model_id => {
@@ -564,12 +621,113 @@ export const getComments = model_id => {
   }));
 };
 
-export const postUpvote = model_id => {
-  const requestString = `${BASE_URL}/upvote`;
+// TEMPLATE WRAPPERS -----------
+export const addTemplate = data => {
+  /**
+   * Adds a template
+   * Returns POST_TEMPLATE_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates`;
   return axios
-    .post(
+    .post(requestString, data, {
+      headers: {
+        "Content-Type": "application/JSON"
+      }
+    })
+    .catch(error => {
+      ({
+        type: "POST_TEMPLATE_FAIL",
+        error
+      });
+    });
+};
+export const deleteTemplate = Template_ID => {
+  /**
+   * Deletes a template
+   * Returns DELETE_TEMPLATE_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates/${Template_ID}`;
+  return axios
+    .delete(requestString, {
+      headers: {}
+    })
+    .catch(error => {
+      ({
+        type: "DELETE_TEMPLATE_FAIL",
+        error
+      });
+    });
+};
+export const getTemplates = () => {
+  /**
+   * Returns all templates
+   * Returns GET_TEMPLATE_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates`;
+  return axios
+    .get(requestString, {
+      headers: {
+        "Content-Type": "application/JSON"
+      }
+    })
+    .catch(error => {
+      ({
+        type: "GET_TEMPLATE_FAIL",
+        error
+      });
+    });
+};
+export const getTemplateByID = Template_ID => {
+  /**
+   * Returns template given ID
+   * Returns GET_TEMPLATE_ID_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates/${Template_ID}`;
+  return axios
+    .get(requestString, {
+      headers: {
+        "Content-Type": "application/JSON",
+        "x-access-token": localStorage.getItem("token")
+      }
+    })
+    .catch(error => {
+      ({
+        type: "GET_TEMPLATE_ID_FAIL",
+        error
+      });
+    });
+};
+
+export const saveTemplate = (data, Template_ID) => {
+  /**
+   * Edits a template
+   * Returns UPDATE_TEMPLATE_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates/${Template_ID}`;
+  return axios
+    .put(requestString, data, {
+      headers: {
+        "Content-Type": "application/JSON",
+        "x-access-token": localStorage.getItem("token")
+      }
+    })
+    .catch(error => {
+      ({
+        type: "UPDATE_TEMPLATE_FAIL",
+        error
+      });
+    });
+};
+export const saveTemplateDraft = (Template_ID, draft) => {
+  /**
+   * Edits a template's draft
+   * Returns UPDATE_TEMPLATE_DRAFT_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates/${Template_ID}/draft`;
+  return axios
+    .put(
       requestString,
-      { upvoteLocation: model_id },
+      { draft },
       {
         headers: {
           "Content-Type": "application/JSON",
@@ -577,10 +735,73 @@ export const postUpvote = model_id => {
         }
       }
     )
-    .catch(error => ({
-      type: "SAVE_UPVOTE_FAIL",
-      error
-    }));
+    .catch(error => {
+      ({
+        type: "SAVE_TEMPLATE_DRAFT_FAIL",
+        error
+      });
+    });
+};
+export const saveTemplateName = (data, Template_ID) => {
+  /**
+   * Edits a template's name
+   * Returns UPDATE_TEMPLATE_NAME_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates/${Template_ID}/name`;
+  return axios
+    .put(requestString, data, {
+      headers: {
+        "Content-Type": "application/JSON"
+      }
+    })
+    .catch(error => {
+      ({
+        type: "UPDATE_TEMPLATE_NAME_FAIL",
+        error
+      });
+    });
+};
+
+export const saveTemplatePhases = (data, Template_ID) => {
+  /**
+   * Edits a template's phases
+   * Returns UPDATE_TEMPLATE_PHASES_FAIL upon failure
+   */
+  const requestString = `${BASE_URL}/templates/${Template_ID}/phases`;
+  return axios
+    .put(requestString, data, {
+      headers: {
+        "Content-Type": "application/JSON"
+      }
+    })
+    .catch(error => {
+      ({
+        type: "UPDATE_TEMPLATE_PHASES_FAIL",
+        error
+      });
+    });
+};
+
+export const postUpvote = async model_id => {
+  const validUser = await checkValidUser();
+  if (validUser) {
+    const requestString = `${BASE_URL}/upvote`;
+    return axios
+      .post(
+        requestString,
+        { upvoteLocation: model_id },
+        {
+          headers: {
+            "Content-Type": "application/JSON",
+            "x-access-token": localStorage.getItem("token")
+          }
+        }
+      )
+      .catch(error => ({
+        type: "SAVE_UPVOTE_FAIL",
+        error
+      }));
+  }
 };
 
 export const getUpvotes = model_id => {
@@ -589,4 +810,21 @@ export const getUpvotes = model_id => {
     type: "GET_UPVOTE_FAIL",
     error
   }));
+};
+
+export const checkToken = () => {
+  const requestString = `${BASE_URL}/auth/checkToken`;
+  return axios
+    .get(requestString, {
+      headers: {
+        "Content-Type": "application/JSON",
+        "x-access-token": localStorage.getItem("token")
+      }
+    })
+    .catch(error => {
+      ({
+        type: "GET_TOKEN_CHECK_FAIL",
+        error
+      });
+    });
 };
